Website Refresh May 2024
I was getting tired of the main design of the site so decided to give it a refresh. Some things may be broken, but most things should work fine. I will get around to anything else that needs done eventually. Thank you for checking it out. Feel free to give me feedback. :)